refactor: move settings to canonical package
