+++ docs/architecture/python-compatibility-shims.md
@@ -60,7 +60,7 @@ import them.
 | `nodes.py` mapped public classes | 18 direct compatibility re-exports plus audited private/residual debt | `easyuse_anima.nodes.*_nodes` | #184 B-04 through B-11, #188 | Existing 0.5.2 surface; B-04 through B-09b2 canonicalized all mapped adapters; B-11 completes the shim | Root mappings and workflows; repository tests are not public-support evidence; no confirmed external direct importer | No scheduled removal; public breaking-change gate after N+1 at earliest |
 | `api.py` route-registration surface | Current implementation; planned API shim | `easyuse_anima.api.router` and `easyuse_anima.api.routes.*` | #165, #186 D-02-D-07 | Existing 0.5.2 surface; convert during D-02-D-07/D-14 | Root entrypoint side-effect import, frontend endpoints, API tests | Unscheduled; N+1 gate and route parity |
 | `api_contract.py` request/error helpers | Phase C temporary implementation; D-02 move and D-14 shim decision pending | `easyuse_anima.api.requests`, `responses`, and `errors` | #165, #186 D-02/D-14 | Introduced by #165; convert in D-02 and freeze any required root shim in D-14 | `api.py`, API contract tests, Registry package-closure test | Unscheduled; internal consumers canonical and contract/package parity pass |
-| `settings.py` | Current implementation; planned settings shim | `easyuse_anima.settings.*` | #163, #186 D-09 | Existing 0.5.2 surface; convert in D-09/D-14 | `api.py`, `nodes.py`, `wildcard_engine.py`, settings tests | Unscheduled; N+1 gate and settings migration/round-trip |
+| `settings.py` | Explicit direct re-export shim (D-09) | `easyuse_anima.settings.schema`, `.repository`, and `.service` | #163, #186 D-09 | Existing 0.5.2 module-owned public surface; exact `__all__` and identity fixture | External/legacy imports and settings compatibility tests; production callers use canonical modules | Unscheduled; first canonical+shim release N not yet recorded, then N+1 gate and settings migration/round-trip |
 | `storage.py` | Explicit direct re-export shim (D-08) | `easyuse_anima.infrastructure.filesystem.atomic_json` and `.paths` | #163, #186 D-08 | Existing 0.5.2 supported module-owned public surface; exact `__all__` and identity fixture | External/legacy imports and storage compatibility tests; production callers use canonical modules | Unscheduled; first canonical+shim release N not yet recorded, then N+1 gate and last-known-good/atomic-write parity |
 | `autocomplete_dataset.py` | Current implementation; planned autocomplete shim | `easyuse_anima.autocomplete.*` | #162, #186 D-11 | Existing 0.5.2 surface; convert in D-11/D-14 | `api.py`, autocomplete/frontend API tests | Unscheduled; N+1 gate and result/ranking/API parity |
 | `wildcard_engine.py` | Current implementation; planned wildcard shim | `easyuse_anima.wildcard.*` | #184, #186 D-12 | Existing 0.5.2 surface; convert in D-12/D-14 | root entrypoint, `nodes.py`, `api.py`, wildcard/workflow tests | Unscheduled; N+1 gate and seed/expansion/workflow parity |
@@ -1376,11 +1376,22 @@ EasyUseAnimaWildcard
 
 ### `settings.py`
 
-- Confirmed current internal consumers use settings load/save/public helpers,
-  long-text helpers, autocomplete/NAIA/metadata/translation resolvers, and
-  translation defaults/types.
-- Canonical target: `easyuse_anima.settings.schema`, `migrations`,
-  `repository`, and `service`.
+- D-09 moves 11 module-owned schema/default/key-map containers to
+  `easyuse_anima.settings.schema`, two import-time file paths and four public
+  file-backed operations to `.repository`, and 22 public projection/resolver
+  functions to `.service`.
+- The root module lists exactly those 39 module-owned public symbols in
+  `__all__` and binds each as the identical canonical object.
+- Raw JSON, long-text normalization, Comfy settings discovery/overlay,
+  string-conversion, and private resolver helpers remain unsupported
+  canonical-owner test seams. Imported filesystem and translation objects are
+  not promoted into the root compatibility surface.
+- Internal production imports use precise canonical owners. Canonical node and
+  prompt modules no longer fall back through the root settings shim.
+- D-09 preserves import-time paths, mutable schema-container identity,
+  call-time store construction, candidate order, precedence, round-trip,
+  clamping, error, and return behavior. It introduces no migration, cache,
+  repository singleton, or RuntimeServices lifecycle.
 - Removal gate: all internal imports are canonical; 0.5.2 settings and long-text
   fixtures migrate and round-trip; original data survives migration/write
   failure; root/canonical supported-symbol identity passes.
