+++ easyuse_anima/nodes/prompt_nodes.py
@@ -18,13 +18,12 @@ from ..prompt.fields import (
     _filter_metadata_prompt,
     _join_prompt_tokens,
 )
+from ..settings.service import resolve_metadata_filter_words
 
 try:
     from ...anima_prompt import correct_prompt, load_knowledge_base
-    from ...settings import resolve_metadata_filter_words
 except ImportError:
     from anima_prompt import correct_prompt, load_knowledge_base
-    from settings import resolve_metadata_filter_words
 
 
 def _correct_prompt_with_report(
