+++ easyuse_anima/settings/__init__.py
@@ -1,0 +1,3 @@
+"""Settings schema, persistence, and service owners."""
+
+__all__ = ()
