+++ tests/test_registry_scanner_safety.py
@@ -70,6 +70,10 @@ EXPECTED_PYTHON_PACKAGE_FILES = {
     "easyuse_anima/profiles/__init__.py",
     "easyuse_anima/profiles/contract.py",
     "easyuse_anima/profiles/mutation.py",
+    "easyuse_anima/settings/__init__.py",
+    "easyuse_anima/settings/repository.py",
+    "easyuse_anima/settings/schema.py",
+    "easyuse_anima/settings/service.py",
     "easyuse_anima/registration.py",
     "easyuse_anima/prompt/__init__.py",
     "easyuse_anima/prompt/correction.py",
