+++ tools/check_python_import_boundaries.py
@@ -42,6 +42,7 @@ EXPECTED_GROUPS = (
     ("lora", 184, "easyuse_anima/lora/", "feature"),
     ("naia", 184, "easyuse_anima/naia/", "feature"),
     ("profiles", 163, "easyuse_anima/profiles/", "feature"),
+    ("settings", 186, "easyuse_anima/settings/", "feature"),
     ("translation", 186, "easyuse_anima/translation/", "feature"),
 )
 ALLOWED_ROLES = frozenset({"common", "feature", "infrastructure"})
